This structure collects count from multiple cred structures.  Of course it
can't use a smaller type.

PR:			283747
Reviewed by:		olce, mjg, markj
Differential Revision:	https://reviews.freebsd.org/D49562
Fixes:			37337709d3334f32650ba3a7c529fa013ed5e1f2

+++ sys/sys/ucred.h
@@ -198,8 +198,8 @@ struct thread;
 
 struct credbatch {
 	struct ucred *cred;
-	int users;
-	int ref;
+	u_int users;
+	long ref;
 };
 
 static inline void
